Move random data to importers (#877)
