+++ .changeset/wild-baboons-matter.md
@@ -1,0 +1,6 @@
+---
+'@xata.io/importer': patch
+'@xata.io/cli': patch
+---
+
+Expose random data command programmatically
